Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/unhandled-promise-rejections/promise-rejection-events.sharedworker-expected.txt

CONSOLE MESSAGE: line 10: ReferenceError: Can't find variable: SharedWorker

FAIL Promise rejection events tests: in a shared worker context ReferenceError: Can't find variable: SharedWorker
